ml-stat: heuristics order for names with comma
Fix order of names with comma, as it is typical for corporate names to be
send as "Surname, FirstName" while the prefered style is
"FirstName Surname". Some people get if wrong and just remove the comma,
in such cases it's better to have the one with comma kept.

I have checked all emails with multiple identities for 6.12...6.13 and
it looks to work great for the mentioned cases, even for cases with more
than two names.

Signed-off-by: Przemek Kitszel <przemyslaw.kitszel@intel.com>